Rectification

+++ Back-end/controller/sauce.js
@@ -23,17 +23,17 @@ exports.modifySauce = (req, res, next) => {
     .then(sauce => {
       const filename = sauce.imageUrl.split('/images/')[1];
       fs.unlink(`images/${filename}`, () => {
-        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
-          .catch(error => res.status(400).json({ error }));
+        if (sauce.userId !== req.auth.userId) {
+          return res.status(403).json({
+            error: new Error("Requête non autorisée !"),
+          });
+        } else {
+          Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce modifiée !' }))
+            .catch(error => res.status(400).json({ error }));
+        }
       });
-      if (sauce.userId !== req.auth.userId) {
-        return response.status(403).json({
-          error: new Error("Requête non autorisée !"),
-        });
-      }
     })
-
 };
 
 exports.deleteSauce = (req, res, next) => {
@@ -41,9 +41,16 @@ exports.deleteSauce = (req, res, next) => {
     .then(sauce => {
       const filename = sauce.imageUrl.split('/images/')[1];
       fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce supprimé !' }))
-          .catch(error => res.status(400).json({ error }));
+        if (sauce.userId !== req.auth.userId) {
+          return res.status(403).json({
+            error: new Error("Requête non autorisée !"),
+          });
+        } else {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Sauce supprimé !' }))
+            .catch(error => res.status(400).json({ error }));
+        }
+
       });
     })
     .catch(error => res.status(500).json({ error }));
@@ -90,7 +97,7 @@ exports.likeSauce = (req, res) => {
 
       }
 
-      res.status(200).json({ message: 'like pris en compte' })
+      res.status(200).json({ message: 'sauce liké' })
     })
     .catch(error => {
       res.status(500).json({ error })
